Create pagerank.py
The input to this implementation is in the file: fullgraph.dat. This is a pickled, gzip file written at the end of running the code for part 1 of this homework assignment.

1) To see the command line options:

python pr.py -h

2) To run the experiment for convergence analysis:

python pr.py -c

This run might take about 40 seconds or more depending on the machine. Outputs the plot in file called prconvergence.pdf

3) To do personalized page rank for a user (eg. estepju):

python pr.py -r estepju

Outputs the plot for percentile in a file called ppr.pdf

4) To merely see the PR values of largest CC:

python pr.py -p

If the graph is large, redirecting the output is recommended for offline analysis.